+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Remove debug
+* Updated translations
 
 ### 1.4.5: May 31th, 2021
 * Automatic refresh of data now executed one minute after cron event has executed, to let time to do the actions